move placebos to their own files
