+++ src/main/java/frc/robot/subsystems/climber/ClimberPlacebo.java
@@ -1,0 +1,23 @@
+package frc.robot.subsystems.climber;
+
+public final class ClimberPlacebo implements ClimberIO {
+
+    @Override
+    public void setRightSpeed(double speed) {
+    }
+
+    @Override
+    public void setLeftSpeed(double speed) {
+    }
+
+    @Override
+    public double getRightMotorCurrent() {
+        return 0;
+    }
+
+    @Override
+    public double getLeftMotorCurrent() {
+        return 0;
+    }
+
+}
