+++ src/main/java/frc/robot/subsystems/indexer/IndexerPlacebo.java
@@ -1,0 +1,14 @@
+package frc.robot.subsystems.indexer;
+
+public final class IndexerPlacebo implements IndexerIO {
+
+    @Override
+    public boolean isBeamBroken() {
+        return false;
+    }
+
+    @Override
+    public void setMotor(double speed) {
+    }
+
+}
