+++ src/main/java/frc/robot/subsystems/intake/IntakeIO.java
@@ -1,13 +1,5 @@
 package frc.robot.subsystems.intake;
 
 public interface IntakeIO {
-    public static final class Placebo implements IntakeIO {
-
-        @Override
-        public void setMotors(double speed) {
-        }
-
-    }
-
     public void setMotors(double speed);
 }
