+++ src/main/java/frc/robot/subsystems/intake/IntakePlacebo.java
@@ -1,0 +1,9 @@
+package frc.robot.subsystems.intake;
+
+public final class IntakePlacebo implements IntakeIO {
+
+    @Override
+    public void setMotors(double speed) {
+    }
+
+}
