+++ src/main/java/frc/robot/subsystems/shooter/ShooterPlacebo.java
@@ -1,0 +1,14 @@
+package frc.robot.subsystems.shooter;
+
+public final class ShooterPlacebo implements ShooterIO {
+
+    @Override
+    public void setMotor(double speed) {
+    }
+
+    @Override
+    public double getEncoderSpeed() {
+        return 0;
+    }
+
+}
